Refactor combobox & multicombobox by unifying typings
