+++ src/components/base/combobox/index.ts
@@ -1,2 +1,3 @@
 export * from "./combobox";
 export * from "./multi-combobox/multi-combobox";
+export * from "./types";
